requirements.txt
Django>=5.2
djangorestframework
python-docx
Pillow
django-cors-headers
django-filter
djangorestframework-simplejwt